pyproject.toml
[project]
name = "foto2pdf"
version = "0.1.0"
description = "A tool for processing and converting photos to PDF"
authors = [
    { name = "Olaf Geibig", email = "olaf@geibig.dev" },
]
dependencies = [
    "pillow>=10.2.0",
    "deskew>=1.3.0",
    "numpy>=1.26.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=foto2pdf --cov-report=term-missing"

[project.scripts]
foto2pdf = "foto2pdf.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.1",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
